+++ hoomd/hpmc/module_simple_polygon.cc
@@ -3,15 +3,11 @@
 
 // Include the defined classes that are to be exported to python
 #include "ComputeFreeVolume.h"
-#include "IntegratorHPMC.h"
 #include "IntegratorHPMCMono.h"
 
 #include "ComputeSDF.h"
 #include "ShapeSimplePolygon.h"
-#include "ShapeUnion.h"
 
-#include "ExternalField.h"
-#include "ExternalFieldHarmonic.h"
 #include "ExternalFieldWall.h"
 
 #include "UpdaterGCA.h"
@@ -38,8 +34,6 @@ void export_simple_polygon(pybind11::module& m)
     export_UpdaterMuVT<ShapeSimplePolygon>(m, "UpdaterMuVTSimplePolygon");
     export_UpdaterGCA<ShapeSimplePolygon>(m, "UpdaterGCASimplePolygon");
 
-    export_ExternalFieldInterface<ShapeSimplePolygon>(m, "ExternalFieldSimplePolygon");
-    export_HarmonicField<ShapeSimplePolygon>(m, "ExternalFieldHarmonicSimplePolygon");
     export_ExternalFieldWall<ShapeSimplePolygon>(m, "WallSimplePolygon");
 
 #ifdef ENABLE_HIP