+++ hoomd/hpmc/module_sphere.cc
@@ -3,16 +3,12 @@
 
 // Include the defined classes that are to be exported to python
 #include "ComputeFreeVolume.h"
-#include "IntegratorHPMC.h"
 #include "IntegratorHPMCMono.h"
 #include "IntegratorHPMCMonoNEC.h"
 
 #include "ComputeSDF.h"
 #include "ShapeSphere.h"
-#include "ShapeUnion.h"
 
-#include "ExternalField.h"
-#include "ExternalFieldHarmonic.h"
 #include "ExternalFieldWall.h"
 
 #include "UpdaterGCA.h"
@@ -40,8 +36,6 @@ void export_sphere(pybind11::module& m)
     export_UpdaterMuVT<ShapeSphere>(m, "UpdaterMuVTSphere");
     export_UpdaterGCA<ShapeSphere>(m, "UpdaterGCASphere");
 
-    export_ExternalFieldInterface<ShapeSphere>(m, "ExternalFieldSphere");
-    export_HarmonicField<ShapeSphere>(m, "ExternalFieldHarmonicSphere");
     export_ExternalFieldWall<ShapeSphere>(m, "WallSphere");
 
 #ifdef ENABLE_HIP