+++ hoomd/hpmc/module_union_sphere.cc
@@ -4,13 +4,10 @@
 // Include the defined classes that are to be exported to python
 #include "ComputeFreeVolume.h"
 #include "ComputeSDF.h"
-#include "IntegratorHPMC.h"
 #include "IntegratorHPMCMono.h"
 
 #include "ShapeUnion.h"
 
-#include "ExternalField.h"
-#include "ExternalFieldHarmonic.h"
 #include "ExternalFieldWall.h"
 
 #include "UpdaterGCA.h"
@@ -37,8 +34,6 @@ void export_union_sphere(pybind11::module& m)
     export_UpdaterMuVT<ShapeUnion<ShapeSphere>>(m, "UpdaterMuVTSphereUnion");
     export_UpdaterGCA<ShapeUnion<ShapeSphere>>(m, "UpdaterGCASphereUnion");
 
-    export_ExternalFieldInterface<ShapeUnion<ShapeSphere>>(m, "ExternalFieldSphereUnion");
-    export_HarmonicField<ShapeUnion<ShapeSphere>>(m, "ExternalFieldHarmonicSphereUnion");
     export_ExternalFieldWall<ShapeUnion<ShapeSphere>>(m, "WallSphereUnion");
 
 #ifdef ENABLE_HIP